* Make IntrospectAsync and InteractAsync methods internal. * Move AuthenticationClient.cs convenience methods to IdxClient.
