+++ src/Okta.Idx.Sdk.IntegrationTests/TestIdxClient.cs
@@ -8,7 +8,7 @@ namespace Okta.Idx.Sdk.IntegrationTests
 {
     public class TestIdxClient
     {
-        public static IIdxClient Create(IdxConfiguration configuration = null)
+        public static IdxClient Create(IdxConfiguration configuration = null)
         {
             // Configuration is expected to be stored in environment variables on the test machine.
             // A few tests pass in a configuration object, but this is just to override and test specific things.
