+++ src/Okta.Idx.Sdk/AuthenticationOptions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Okta.Idx.Sdk
+{
+    public class AuthenticationOptions
+    {
+        public string Username { get; set; }
+
+        public string Password { get; set; }
+    }
+}
