+++ src/Okta.Idx.Sdk/ChangePasswordOptions.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Okta.Idx.Sdk
+{
+    public class ChangePasswordOptions
+    {
+        public string NewPassword { get; set; }
+    }
+}
